requirements.txt
absl-py==1.4.0
anyio==3.7.0
array-record==0.2.0
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
Cython==0.29.35
dataclasses==0.6
decorator==5.1.1
dm-tree==0.1.8
etils==1.3.0
exceptiongroup==1.1.1
fastapi==0.97.0
fire==0.5.0
flatbuffers==23.5.26
gast==0.5.4
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.11.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.48.2
h11==0.14.0
h5py==3.8.0
httplib2==0.22.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kaggle==1.5.13
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
librosa==0.8.1
llvmlite==0.36.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.4.3
neural-structured-learning==1.4.0
numba==0.53.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==20.9
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.3
pooch==1.7.0
promise==2.3
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.9
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scann==1.2.6
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
starlette==0.27.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.8.0
tensorflow-gpu==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.0
tensorflow-model-optimization==0.7.5
tensorflowjs==3.18.0
termcolor==2.3.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.3.0
tf-slim==1.1.0
tflite-model-maker==0.4.2
tflite-support==0.4.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.65.0
typeguard==2.13.3
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.25.11
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
absl-py==1.4.0
anyio==3.7.0
array-record==0.2.0
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==42.0.7
cycler==0.11.0
Cython==0.29.35
dataclasses==0.6
decorator==5.1.1
dm-tree==0.1.8
etils==1.3.0
exceptiongroup==1.1.1
fastapi==0.97.0
fire==0.5.0
flatbuffers==23.5.26
gast==0.5.4
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.11.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.48.2
h11==0.14.0
h5py==3.8.0
httplib2==0.22.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jwt==1.3.1
kaggle==1.5.13
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
librosa==0.8.1
llvmlite==0.36.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.4.3
neural-structured-learning==1.4.0
numba==0.53.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==20.9
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.3
pooch==1.7.0
promise==2.3
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.9
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scann==1.2.6
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
starlette==0.27.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.8.0
tensorflow-gpu==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.0
tensorflow-model-optimization==0.7.5
tensorflowjs==3.18.0
termcolor==2.3.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.3.0
tf-slim==1.1.0
tflite-model-maker==0.4.2
tflite-support==0.4.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.65.0
typeguard==2.13.3
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.25.11
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
